Updated E2E test to include back link journeys
